Develop

+++ api/src/main/java/com/epam/esm/controller/OrderController.java
@@ -1,14 +1,17 @@
 package com.epam.esm.controller;
 
 import com.epam.esm.dto.GiftCertificateDto;
-import com.epam.esm.dto.OrderDto;
+import com.epam.esm.dto.ReceiveOrderDto;
+import com.epam.esm.dto.ResponseOrderDto;
 import com.epam.esm.service.OrderService;
+import com.epam.esm.service.SecurityService;
 import com.epam.esm.service.hateoas.HateoasService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.annotation.security.RolesAllowed;
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
 
@@ -21,17 +24,20 @@ import static org.springframework.http.HttpStatus.CREATED;
 public class OrderController {
     private final OrderService orderService;
     private final HateoasService hateoasService;
+    private final SecurityService securityService;
 
     /**
      * The method allows {@link com.epam.esm.entity.Order} creating.
      *
-     * @param orderDto should be valid according to {@link OrderDto}. Otherwise, certificate won`t be created. Error
-     *                 will be returned(400).
+     * @param orderDto should be valid according to {@link ResponseOrderDto}. Otherwise, certificate won`t be created.
+     *                 Error will be returned(400).
      * @return ResponseEntity witch contains created order with generated id. Response code 201.
      */
     @PostMapping
-    public ResponseEntity<OrderDto> createOrder(@Valid @RequestBody OrderDto orderDto) {
-        OrderDto createdOrderDto = orderService.createOrder(orderDto);
+    @RolesAllowed({"ADMIN", "USER"})
+    public ResponseEntity<ResponseOrderDto> createOrder(@Valid @RequestBody ReceiveOrderDto orderDto) {
+        securityService.validateUserAccess(orderDto.getUserId());
+        ResponseOrderDto createdOrderDto = orderService.createOrder(orderDto);
         hateoasService.attachHateoas(createdOrderDto);
         return ResponseEntity.status(CREATED).body(createdOrderDto);
     }
@@ -39,12 +45,15 @@ public class OrderController {
     /**
      * Provides ability to update order entirely.
      *
-     * @param orderDto should have all fields filled and valid. Otherwise error will be returned(400).
-     * @return updated order wrapped into {@link OrderDto}. Response code 200.
+     * @return updated order wrapped into {@link ResponseOrderDto}. Response code 200.
      */
-    @PutMapping
-    public ResponseEntity<OrderDto> updateOrder(@Valid @RequestBody OrderDto orderDto) {
-        OrderDto updatedOrderDto = orderService.updateOrder(orderDto);
+    @PutMapping("/{orderId}")
+    @RolesAllowed({"ADMIN", "USER"})
+    public ResponseEntity<ResponseOrderDto> updateOrder(@Valid @RequestBody ReceiveOrderDto receiveOrderDto,
+                                                        @PathVariable("orderId") @Min(1) Long orderId) {
+        ResponseOrderDto responseOrderDto = orderService.getOrderById(orderId);
+        securityService.validateUserAccess(responseOrderDto.getUser().getId());
+        ResponseOrderDto updatedOrderDto = orderService.updateOrder(receiveOrderDto, orderId);
         hateoasService.attachHateoas(updatedOrderDto);
         return ResponseEntity.ok().body(updatedOrderDto);
     }
@@ -56,7 +65,10 @@ public class OrderController {
      * @return responseEntity having empty body. Response code 204.
      */
     @DeleteMapping("/{orderId}")
+    @RolesAllowed({"ADMIN", "USER"})
     public ResponseEntity<GiftCertificateDto> deleteOrder(@PathVariable("orderId") @Min(1) Long orderId) {
+        ResponseOrderDto orderDto = orderService.getOrderById(orderId);
+        securityService.validateUserAccess(orderDto.getUser().getId());
         orderService.deleteOrder(orderId);
         return ResponseEntity.noContent().build();
     }
